requirements.txt
Django==5.1
djangorestframework==3.14.0
ruff==0.3.0
python-decouple==3.8
pylint-django==2.5.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
validate-docbr==1.10.0
Faker==28.0.0
